feat: update CaixaService, EstoqueService, and ServicoService to use BigDecimal for financial calculations and adjust related methods

+++ src/main/java/com/br/fiap/oficina/service/ServicoService.java
@@ -3,6 +3,7 @@ import com.br.fiap.oficina.model.entity.Servico;
 import com.br.fiap.oficina.model.repository.ServicoRepository;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -22,7 +23,7 @@ public class ServicoService {
         return (List<Servico>) repository.findAll();
     }
 
-    public void cadastrarServico(String nome, String descricao, Double preco, Integer duracao) {
+    public void cadastrarServico(String nome, String descricao, BigDecimal preco, Integer duracao) {
         repository.save(Servico.builder()
                 .nome(nome)
                 .descricao(descricao)
@@ -36,7 +37,7 @@ public class ServicoService {
         repository.delete(servico);
     }
 
-    public void atualizarServico(Long id, String nome, String descricao, Double preco, Integer duracao) {
+    public void atualizarServico(Long id, String nome, String descricao, BigDecimal preco, Integer duracao) {
         var servico = repository.findById(id).orElseThrow();
         servico.setNome(nome);
         servico.setDescricao(descricao);
